backend/requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
Pillow==10.0.1
pytesseract==0.3.10
torch>=2.0.0
ultralytics==8.0.134
gunicorn==21.2.0
